% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{callback}
\alias{callback}
\title{callback function for readr read_delim_chunked in sql_upload_txt_chunked()}
\usage{
callback(chunk, pos)
}
\arguments{
\item{chunk}{chunk of txt file loaded by readr::read_delim_chunked}

\item{pos}{position of chunk loaded by readr::read_delim_chunked}
}
\value{
TRUE invisibly also writes a table in a SQL database and saves out a copy of chunk in parquet
}
\description{
Not usually called do not run unless re-developing
A function to be run by readr inside of sql_upload_txt_chunked() or
sql_upload_csv_chunked() (referred to collectively) this function writes the chunk of the file
read by readr to a SQL table specified in sql_upload_txt_chunked()
and produces a parquet file backup of each chunk read by readr, combined by combine_parquet_backup
called in sql_upload_parquet_backup_from_txt. Parquet writing can disabled by
specifying write_parquet = FALSE in sql_upload_txt_chunked
#' Upload txt file to staging table and create parquet backup
}
\note{
getOption() provides additional params set by sql_upload_txt_chunked/csv
}
\examples{
f <- function(x,pos) print(pos)
readr::read_csv_chunked(readr::readr_example("mtcars.csv"),
callback = readr::SideEffectChunkCallback$new(f),
chunk_size = 10
)
}
\author{
Owen Pullen
}
