% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_tracker_clear.R
\name{case_tracker_clear}
\alias{case_tracker_clear}
\title{Clear the case list and case count tracking tables}
\usage{
case_tracker_clear()
}
\value{
None. Resets the \code{tracking_list}, \code{tracking_count}, and \code{last_status}.
}
\description{
Clears the \code{tracking_list} and \code{tracking_count} in the environment, effectively
resetting the tracking data to start fresh for a new pipeline run or region.
}
