% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_tracker_save.R
\name{case_tracker_save}
\alias{case_tracker_save}
\title{Save the tracking list to a CSV file}
\usage{
case_tracker_save(file_path)
}
\arguments{
\item{file_path}{Character. File path where the tracking list CSV should
be saved.}
}
\value{
None. Saves the tracking list to a CSV file at the specified
\code{file_path}.
}
\description{
Saves the current tracking list to a specified file path as a CSV file.
}
