% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_tracker_update.R
\name{case_tracker_update}
\alias{case_tracker_update}
\title{Update the case status based on the current case list at a checkpoint}
\usage{
case_tracker_update(cases, checkpoint, type = "case_list")
}
\arguments{
\item{cases}{A character vector of the current case IDs still in the pipeline.}

\item{checkpoint}{Character. The name of the checkpoint in the data pipeline.}

\item{type}{Character. Indicates whether to update \code{"case_list"} (default)
or \code{"case_count"}.}
}
\value{
None. Updates the status of cases to "active" or "Dropped at \if{html}{\out{<checkpoint>}}".
}
\description{
Updates the case status in the tracking list based on the provided list of
current case IDs at a specific checkpoint in the pipeline.
}
