% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{clean_alpha}
\alias{clean_alpha}
\title{Clean alpha strings (letters only)}
\usage{
clean_alpha(str, remove_punct = FALSE, remove_all = FALSE)
}
\arguments{
\item{str}{string to be cleaned}

\item{remove_punct}{\itemize{\item boolean TRUE/ FALSE, TRUE removes all punctuation,
\item when FALSE will remove punctuation that is not \verb{!&',-.:;?}
\item This argument is default FALSE
}}

\item{remove_all}{boolean \code{TRUE/FALSE}, TRUE removes anything non A-Z or a-z including spaces and punctuation}
}
\value{
character string
}
\description{
Wrapper around a regex to remove all non letter characters, will remove all numbers from strings
\itemize{
\item remove_punct = FALSE \verb{[^A-Za-z !&',-.:;?]|[0-9]} will remove all punctuation that\if{html}{\out{<br>}}
is not \verb{!&',-.:;?} will not remove spaces
\item remove_punct = TRUE \verb{[^A-za-z ]} will remove all non A-z characters except spaces
\item remove_all = TRUE \verb{[^[:alpha:]]} will remove all non A-z characters including spaces and all punctuation
}
If you are interested visit: https://regex101.com to test and build regexs
}
\examples{
clean_alpha("Test, string! ;", remove_punct = TRUE, remove_all = FALSE)
clean_alpha("Test, string! ;", remove_punct = FALSE, remove_all = FALSE)
# Argument remove_all is best used on ID variables only as it removes all spaces
clean_alpha("Test, string! ;", remove_punct = FALSE, remove_all = TRUE)
}
\seealso{
\code{\link[=clean_alpha_numeric]{clean_alpha_numeric()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}} \code{\link[=remove_whitespace]{remove_whitespace()}}
}
\author{
Owen Pullen
}
