% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{clean_alpha_numeric}
\alias{clean_alpha_numeric}
\title{Clean alpha numeric strings (letters and numbers only)}
\usage{
clean_alpha_numeric(
  str,
  remove_punct = FALSE,
  remove_all = FALSE,
  alpha_num_id = FALSE
)
}
\arguments{
\item{str}{string to be cleaned}

\item{remove_punct}{boolean TRUE/FALSE, remove punctuation from string when TRUE all punctuation will be removed,
when FALSE punctuation that is not \verb{!&',-.:;?} will be removed. Default FALSE}

\item{remove_all}{boolean TRUE/FALSE, removes all non A-z or 0-9 characters inducing spaces and punctuation,
remove_all will override other boolean arguments}

\item{alpha_num_id}{boolean TRUE/FALSE removes all non A-z characters except '-', including space
use on alphanumeric ID variables only}
}
\value{
character string
}
\description{
Wrapper around regex to remove all non alpha-numeric characters
\itemize{
\item remove_punct = TRUE regex is \verb{[^[:alnum:]]} will remove all non A-z and 0-9 characters
\item remove_punct = FALSE regex is \verb{[^A-Za-z0-9!&',-.:;? ]} will remove all non A-z and 0-9 characters
\item remove_all = TRUE regex is \verb{[^[:alnum:]]} will remove all non-alpha numeric characters
\item alpha_num_id = TRUE regex is \verb{[^A-Za-z0-9-]} will remove all non alphanumeric characters expect \code{-}
\item If you are using one a Abc-123-Def-456 type ID and want to remove numeric all other characters use alpha_num_id = TRUE
}
If you are interested visit: https://regex101.com to test and build regexs
}
\examples{
clean_alpha_numeric("ABC-123 ;", remove_punct = FALSE, remove_all = FALSE)
clean_alpha_numeric("ABC-123 ;", remove_punct = TRUE, remove_all = FALSE)
clean_alpha_numeric("ABC-123 ;", remove_punct = FALSE, remove_all = TRUE)
clean_alpha_numeric("Abc-123;'#?>-D<ef-456", alpha_num_id = TRUE)
}
\seealso{
\code{\link[=clean_alpha]{clean_alpha()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}} \code{\link[=remove_whitespace]{remove_whitespace()}}
}
\author{
Owen Pullen
}
