% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sql.R
\name{clean_sql}
\alias{clean_sql}
\title{Clean and Read a SQL query

A utility function to read in a SQL query from a character object, clipboard
or text file and remove all comments for use with database query packages}
\usage{
clean_sql(sql)
}
\arguments{
\item{sql}{a SQL file or text string}
}
\value{
a cleaned SQL query without comments as a character string
}
\description{
Clean and Read a SQL query

A utility function to read in a SQL query from a character object, clipboard
or text file and remove all comments for use with database query packages
}
\examples{
testSQL <- c(
"/*********",
"INTRO HEADER COMMENTS",
"*********/","  SELECT ",
"  [VAR 1]  -- with comments",
",[VAR 2]",",[VAR 3]","FROM DATASET ",
"-- output here")
clean_sql(testSQL)

}
