% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_parquet_files.R
\name{combine_parquet_files}
\alias{combine_parquet_files}
\title{Combines parquet file chunk datasets saved across multiple parquet files}
\usage{
combine_parquet_files(path, pattern, backup_name, remove_old = FALSE)
}
\arguments{
\item{path}{Reads parquet files, combines them into one single parquet file and removes old files}

\item{pattern}{pattern to look for in parquet files to combine
e.g. pattern = "temp" - looks for files with temp in the file name}

\item{backup_name}{name of file this is a string name that the file name will be not including file path
or including file extension e.g. backup_name = "example_file" is for "blah/example_file.parquet"}

\item{remove_old}{removes chunked parquet files written by callback function}
}
\value{
invisibly character vector of parquet files that were merged and a single parquet file (saved/written)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Combine parquet files into larger parquet files - useful for when writing parquet files in chunks
then later combining chunks into single file
The ordering may fail with a linux OS
}
\examples{
t <- tempdir(check = TRUE)
t1 <- paste0(t, "\\\\")
iris_1 <- dplyr::filter(iris, Species == "setosa")
iris_2 <- dplyr::filter(iris, Species == "virginica")
arrow::write_parquet(iris_1, paste0(t1, "input_1.parquet"))
arrow::write_parquet(iris_2, paste0(t1, "input_2.parquet"))
combine_parquet_files(path = t1,
                      pattern = "input",
                      backup_name = "sample_of_iris",
                      remove_old = TRUE
                      )

}
\author{
Owen Pullen
}
