% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_from_zip.R
\name{csv_from_zip}
\alias{csv_from_zip}
\title{Download a csv from a zip}
\usage{
csv_from_zip(x)
}
\arguments{
\item{x}{a zip file from the web}
}
\value{
a zip file for ingestion into your chosen readr
}
\description{
Download a csv from a zip
}
\examples{
head(read.csv(csv_from_zip("https://files.digital.nhs.uk/assets/ods/current/succarc.zip")))

}
