% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidm-package.R
\docType{package}
\name{epidm-package}
\alias{epidm}
\alias{epidm-package}
\title{epidm: UK Epidemiological Data Management}
\description{
Contains utilities and functions for the cleaning, processing and management of patient level public health data for surveillance and analysis held by the UK Health Security Agency, UKHSA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexbhatt/epidm}
  \item \url{https://alexbhatt.github.io/epidm/}
  \item Report bugs at \url{https://github.com/alexbhatt/epidm/issues}
}

}
\author{
\strong{Maintainer}: Alex Bhattacharya \email{alex.bhatt@gmail.com} (\href{https://orcid.org/0000-0003-3000-2771}{ORCID})

}
\keyword{internal}
