% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql_file.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{Function to execute SQL files}
\usage{
execute_sql_file(con, query_path)
}
\arguments{
\item{con}{DBI connection object}

\item{query_path}{file path to a sql query, this should be saved in a .sql file
for readability and maintainability}
}
\value{
number of rows affected by query provided
}
\description{
Executes upsert command in SQL database, this will execute any SQL query provided in
a .sql file, so use sensibly. Do not execute SQL files without knowing the expected output.
Users will have to have relevant permissions in the database e.g. if user wants to execute
an upsert (INSERT and UPDATE) user must have SQL insert and update permission in database

This function just provides a simplified way to execute a .sql file handling the reading
into R for the user.
This is similar to \code{\link[=dbExecute]{dbExecute()}} if you have a string query use \code{\link[=dbExecute]{dbExecute()}}
}
