% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsnumber_testing_internals.R
\name{gen_test_nhs}
\alias{gen_test_nhs}
\title{gen test nhs number}
\usage{
gen_test_nhs(n = 1)
}
\arguments{
\item{n}{number of times to be replicated}
}
\description{
gen test nhs number
}
\keyword{internal}
