% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsnumber_testing_internals.R
\name{gen_testing_nhsno_single}
\alias{gen_testing_nhsno_single}
\title{gen test nhs number for single number only called in gen test nhs_number}
\usage{
gen_testing_nhsno_single()
}
\description{
gen test nhs number for single number only called in gen test nhs_number
}
\keyword{internal}
