% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_inpatient_admission_method}
\alias{group_inpatient_admission_method}
\title{Inpatient admission methods}
\format{
\describe{
\item{code}{the admission_method code}
\item{admission_method}{the admission_method grouping as a human readable string}
}
}
\usage{
group_inpatient_admission_method
}
\description{
In order to group hospital inpatient admissions into human readable groups,
a lookup table has been created. These work with Hospital Episode Statistics
(HES) and Secondary Use Services (SUS) data with the admission_method fields.
}
\keyword{datasets}
