% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_episode.R
\name{infection_episode}
\alias{infection_episode}
\title{Patient Episode Groupings

assigns isolates of same species to an episode window
works on line listings only
designed for SGSS data
MONO EPISODES ARE PER ORGANISM PER PATIENT
Note that first specimen date is day zero.}
\usage{
infection_episode(x, group_vars, specimen_date, episode_window = c(0:Inf))
}
\arguments{
\item{group_vars}{a vector containing any columns used to identify the patient, infection, and site}

\item{episode_window}{an integer to specify the length of the episode}

\item{data}{a data.frame or tibble containing the line list}

\item{spec_date}{the column as a date variable containing the specimen date}

\item{episode_type}{a choice of static or rolling for how the window is calculated}
}
\value{
A new dataframe containing the original data, and columns specifying monomicrobial or polymicrobial patient episodes.
}
\description{
Patient Episode Groupings

assigns isolates of same species to an episode window
works on line listings only
designed for SGSS data
MONO EPISODES ARE PER ORGANISM PER PATIENT
Note that first specimen date is day zero.
}
\examples{
episode_test <- structure(
  list(
    pat_id = c(1L, 1L, 1L, 1L, 2L, 2L, 2L),
    sp_date = structure(c(18262, 18263, 18281, 18282, 18262, 18263, 18281),
                        class = "Date"),
    species = c("E. coli", "E. coli", "E. coli", "E. coli", "E. coli", "E. coli","E. coli"),
    spec_type = c("Blood", "Blood", "Blood", "Blood", "Blood", "Blood","Blood")
  ),
  row.names = c(NA, -7L), class = "data.frame")

infection_episode(
  x = episode_test,
  group_vars = c("pat_id", "spec_type", "species")
  specimen_date = "sp_date",
  episode_window = 14
)

}
