% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_recode.R
\name{lookup_recode}
\alias{lookup_recode}
\title{Lookup table switch handler}
\usage{
lookup_recode(col, type = c("species", "specimen", "manual"), .import = NULL)
}
\arguments{
\item{col}{a vector containing the variable to be created/changed}

\item{type}{a character value to denote the lookup}

\item{.import}{a list  in the order list(new,old) containing the
values for another lookup table existing in the environment}
}
\value{
a vector/list object of the recoded field
}
\description{
A function to call an epidm lookup table and
recode  where we are aware of a new value.
}
\details{
Built in are  the organism re-classifications and specimen_type groupings
and a manual mode.
}
\examples{
respec_example <- data.frame(
  sp=respeciate_organism$previous_organism_name,
  ge=gsub("([A-Za-z]+).*", "\\\\1",respeciate_organism$previous_organism_name)
  )
respec_example$new <- lookup_recode("species",respec_example$sp)
head(respec_example)

lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
}
