% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_episode.R
\name{monomicrobial_episode}
\alias{monomicrobial_episode}
\title{Patient Episode Groupings

assigns isolates of same species to an episode window
works on line listings only
designed for SGSS data
MONO EPISODES ARE PER ORGANISM PER PATIENT
Note that first specimen date is day zero.}
\usage{
monomicrobial_episode(
  data,
  patient_id,
  spec_date,
  species,
  spec_type,
  episode_window = c(1:365),
  episode_type = c("static", "rolling")
)
}
\arguments{
\item{data}{a data.frame or tibble containing the line list}

\item{patient_id}{the column as a character containing the unique patient id}

\item{spec_date}{the column as a date variable containing the specimen date}

\item{species}{the column as a character containing the species name}

\item{spec_type}{the column as a character containing the specimen type}

\item{episode_window}{an integer to specify the length of the episode}

\item{episode_type}{a choice of static or rolling for how the window is calculated}
}
\value{
A new dataframe containing the original data, and columns specifying monomicrobial or polymicrobial patient episodes.
}
\description{
Patient Episode Groupings

assigns isolates of same species to an episode window
works on line listings only
designed for SGSS data
MONO EPISODES ARE PER ORGANISM PER PATIENT
Note that first specimen date is day zero.
}
\examples{
dat <- structure(list(pat_id = c(1L, 1L, 1L, 1L, 2L, 2L, 2L), 
sp_date = structure(c(18262, 18263, 18281, 18282, 18262, 18263, 18281), 
  class = "Date"), 
species = c("E. coli", "E. coli", "E. coli", "E. coli", "E. coli", "E. coli", 
  "E. coli"), 
  spec_type = c("Blood", "Blood", "Blood", "Blood", "Blood", "Blood", 
  "Blood")), row.names = c(NA, -7L), class = "data.frame")
  
monomicrobial_episode(data = dat, patient_id = "pat_id", spec_date = "sp_date",
  spec_type = "spec_type", species = "species", episode_window = 14, 
  episode_type = "static")


}
