% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_nhs.R
\name{range_nhs_number}
\alias{range_nhs_number}
\title{range_nhs_number}
\usage{
range_nhs_number(value)
}
\arguments{
\item{value}{an NHS number for validation using range check}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Checks NHS number range against valid range for England, Scotland and Northern Ireland\cr
Examples are not real NHS numbers please do not use real
NHS numbers, in tests, code or documentation
}
\examples{
range_nhs_number("1")
range_nhs_number(200000000000)
range_nhs_number(1111111111)
}
\seealso{
\code{\link[=valid_nhs]{valid_nhs()}} \code{\link[=checksum_nhs_number]{checksum_nhs_number()}} \code{\link[=seq_nhs_number]{seq_nhs_number()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}}
}
\author{
Owen Pullen
}
