% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{remove_non_numeric_char}
\alias{remove_non_numeric_char}
\title{Remove non numeric characters}
\usage{
remove_non_numeric_char(str)
}
\arguments{
\item{str}{string to be cleaned}
}
\value{
character string
}
\description{
Wrapper around the \verb{[^0-9]} regex to remove all non-numeric characters from a string
\itemize{
\item This is particularly useful for cleaning out numeric ID fields.
\item This should not be used where characters are expected
\item This function is useful for cleaning out all numeric Id's such as NHS number
}
}
\examples{
remove_non_numeric_char('123 and 456')
remove_non_numeric_char("12;3*456:78.'9")
}
\seealso{
\code{\link[=clean_alpha_numeric]{clean_alpha_numeric()}} \code{\link[=clean_alpha]{clean_alpha()}} \code{\link[=remove_whitespace]{remove_whitespace()}}
}
\author{
OWen Pullen
}
