% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove all white space characters}
\usage{
remove_whitespace(str)
}
\arguments{
\item{str}{string to be cleaned}
}
\value{
character string
}
\description{
\itemize{
\item Wrapper around \verb{[[:space:]]} regex to remove all white space characters including space tab and new line characters
}
If you are interested visit: https://regex101.com to test and build regexs
}
\examples{
remove_whitespace("A B C")
remove_whitespace("1 2 3")
remove_whitespace(". . .\t")
}
\seealso{
\code{\link[=clean_alpha_numeric]{clean_alpha_numeric()}} \code{\link[=clean_alpha]{clean_alpha()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}}
}
\author{
Owen Pullen
}
