% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respec_organism.R
\name{respec_org}
\alias{respec_org}
\title{Respeciate organism}
\usage{
respec_org(species)
}
\arguments{
\item{species}{a vector containing the organism species name}
}
\value{
a vector containing the updated scientific name for the organism
}
\description{
A function to call the respeciated_organism lookup table and
recode organisms where we are aware of a new classification
}
\examples{
respec_example <- data.frame(
  sp=respeciate_organism$previous_organism_name,
  ge=gsub("([A-Za-z]+).*", "\\\\1",respeciate_organism$previous_organism_name)
  )
respec_example$new <- respec_org(respec_example$sp)
head(respec_example)
}
