% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeciate_generic.R
\name{respeciate_generic}
\alias{respeciate_generic}
\title{Respeciate unspecified samples}
\usage{
respeciate_generic(x, group_vars, species, date, window = c(0:Inf))
}
\arguments{
\item{x}{a data.frame or data.table object}

\item{group_vars}{the minimum grouping set of variables for like samples in a character vector; suggest c('patient_id','specimen_type','genus')}

\item{species}{a character containing the column with the organism species name}

\item{date}{a character containing the column with the specimen/sample date}

\item{window}{an integer representing the number of days for which you will allow a sample to be respeciated}
}
\value{
a data.table with a recharacterised specimen column
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Some samples within SGSS are submitted by laboratories as "GENUS SP" or
"GENUS UNNAMED". However, they may also have a fully identified sample taken
from the same site within a recent time period.  This function captures
species  from another sample within X-days of an unspeciated isolate.
}
\examples{
df <- data.frame(
ptid = c(round(runif(25,1,5))),
spec = sample(c("KLEBSIELLA SP",
                "KLEBSIELLA UNNAMED",
                "KLEBSIELLA PNEUMONIAE",
                "KLEBEIELLA OXYTOCA"),
              25,replace = T),
type = "BLOOD",
date = sample(seq.Date(Sys.Date()-21,Sys.Date(),"day"),25,replace = T)
)

respeciate_generic(x=df,
                   group_vars=c('ptid','type'),
                   species='spec',
                   date='date',
                   window = 14)[]
}
