% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respeciate_organism}
\alias{respeciate_organism}
\title{Respeciated organisms}
\format{
\describe{
\item{previous_organism_name}{What the organism used to be known as, in the form GENUS SPECIES}
\item{organism_species_name}{What the organism is known as now, in the form GENUS SPECIES}
\item{organism_genus_name}{The genus of the recoded organism}
\item{genus_change}{A 0/1 flag to indicate if the genus has changed}
\item{genu_all_species}{A 0/1 flag to indicate if all species under that genus should change}
}
}
\usage{
respeciate_organism
}
\description{
Occasionally, research shows that two organisms, previously thought to be
different are in fact one and the same. The reverse is also true.
This is  a manually updated list.
If there are organisms missing, or new respeciates to be added,
please raise and issue or push request on the
\href{https://github.com/alexbhatt/epidm}{epidm GitHub}
}
\keyword{datasets}
