% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specimen_type_grouping}
\alias{specimen_type_grouping}
\title{Specimen type grouping}
\format{
\describe{
\item{specimen_type}{The primary specimen type with detail}
\item{specimen_group}{A simple grouping of like specimen sites}
}
}
\usage{
specimen_type_grouping
}
\description{
In order to help clean up an analysis based on a group of specimen types,
a lookup table has been created to help group sampling sites.
This is  a manually updated list.
If there are organisms missing, or new respeciates to be added,
please raise and issue or push request on the
\href{https://github.com/alexbhatt/epidm}{epidm GitHub}
}
\keyword{datasets}
