% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_clean.R
\name{sql_clean}
\alias{sql_clean}
\title{Clean and Read a SQL query}
\usage{
sql_clean(sql)
}
\arguments{
\item{sql}{a SQL file or text string}
}
\value{
a cleaned SQL query without comments as a character string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A utility function to read in a SQL query from a character object, clipboard
or text file and remove all comments for use with database query packages
}
\examples{
testSQL <- c(
"/********* INTRO HEADER COMMENTS",
"*********/",
"  SELECT ",
"  [VAR 1]  -- with comments",
",[VAR 2]",",[VAR 3]",
"FROM DATASET ","-- output here")
sql_clean(testSQL)


}
