% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_id.R
\name{uk_patient_id}
\alias{uk_patient_id}
\title{Patient ID record grouping}
\usage{
uk_patient_id(
  x,
  nhs_number,
  hospital_number,
  date_of_birth,
  sex_mfu,
  forename = "NONAME",
  surname = "NONAME",
  .sortOrder,
  .forceCopy = FALSE
)
}
\arguments{
\item{x}{a data.frame or data.table containing the cleaned line list}

\item{nhs_number}{a column as a character containing the patient NHS numbers}

\item{hospital_number}{a column as a character containing the patient
Hospital numbers}

\item{date_of_birth}{a column as a date variable containing the patient
date of birth in date format}

\item{sex_mfu}{column as a character containing the patient sex;
NOTE only works if coded only as character versions of Male/Female/Unknown;
does not currently work with additional options \verb{#future update}}

\item{forename}{a column as a character containing the patient forename;
leave as NONAME if unavailable}

\item{surname}{a column as a character containing the patient surname;
leave as NONAME if unavailable}

\item{.sortOrder}{optional; a column as a character to allow a sorting
order on the id generation}

\item{.forceCopy}{default FALSE; TRUE will force data.table to take a copy
instead of editing the data without reference}
}
\value{
A dataframe with one new variable:
\describe{
\item{\code{id}}{a unique patient id}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Groups patient records from multiple isolates with a single integer patientID
by grouping patient identifiers.

Grouping is based on five stages:
\enumerate{
\item matching nhs number and date of birth
\item Hospital number &  Date of Birth
\item NHS number & Hospital Number
\item Date of Birth & Surname IF nhs unknown
\item Sex & Date of Birth & Fuzzy Name
}

Identifiers are copied over where they are missing or invalid to the grouped
records.
}
