% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsnumber_testing_internals.R
\name{valid_nhs_internal}
\alias{valid_nhs_internal}
\title{alternate method nhs number validation}
\usage{
valid_nhs_internal(nhs_number)
}
\description{
alternate method nhs number validation
}
\keyword{internal}
