% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nhs_number.R
\name{validate_nhs_number}
\alias{validate_nhs_number}
\title{valid_nhs}
\usage{
validate_nhs_number(nhs_number, return_nhsno = FALSE)
}
\arguments{
\item{nhs_number}{an NHS number as a character string, interger or vector.}

\item{return_nhsno}{boolean \link{TRUE}/\link{FALSE} TRUE returns the NHS number as entered if valid and NA if invalid
FALSE returns a 0/1}
}
\description{
NHS Number validation function calls checksum_nhs_number() which validates NHS numbers using the NHS checksum validation algorithm\cr
Also calls seq_nhs_number() which looks for 10 character single number sequences using regex e.g. 9999999999\cr
Calls range_nhs_number() checks NHS number is within the valid range for England, Wales, Scotland and Northern Ireland\cr
Calls remove_non_numeric_char() removes all non numeric characters from a string field
If you want to use the NHS checksum algorithm but do not want to do any other validation use checksum_nhs_number()\cr
Examples are not real NHS numbers please do not use real
NHS numbers, in tests, code or documentation
}
\examples{
valid_nhs("1234567890")
test_vect <- nhsnumbergenerator::generate_nhs_number(100)
valid_nhs(test_vect)
valid_nhs(NA)
}
\seealso{
\code{\link[=range_nhs_number]{range_nhs_number()}} \code{\link[=seq_nhs_number]{seq_nhs_number()}} \code{\link[=checksum_nhs_number]{checksum_nhs_number()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}}
}
\author{
Owen Pullen
}
