% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_tracker_update.R
\name{case_tracker_update}
\alias{case_tracker_update}
\title{Update the status of specified cases in the tracking list}
\usage{
case_tracker_update(cases, new_status)
}
\arguments{
\item{cases}{Character vector of case IDs or data frame with a \code{case_id}
column.}

\item{new_status}{Character. Status message indicating the new state
or reason for exclusion.}
}
\value{
None. Updates the status of specified cases in the tracking list.
}
\description{
Updates the status of cases in the tracking list by their \code{case_id}.
}
