% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nhs_number.R
\name{checksum_nhs_number}
\alias{checksum_nhs_number}
\title{checksum_nhs_number}
\usage{
checksum_nhs_number(value)
}
\arguments{
\item{value}{an NHS number as an integer or character string or vector for validation using
the NHS designed checksum algorithm}
}
\description{
Preforms the checksum validation algorithm  on an NHS number. Can also be called inside a mutate\cr
If you want to use the NHS checksum algorithm but do not want to do any other validation use checksum_nhs_number()\cr
Examples are not real NHS numbers please do not use real
NHS numbers, in tests, code or documentation
}
\examples{
checksum_nhs_number("1234567890")
example = data.frame(example_1 = nhsnumbergenerator::generate_nhs_number(100))
dplyr::mutate(example, checksum_nhs_number(example_1))
}
\seealso{
\code{\link[=range_nhs_number]{range_nhs_number()}} \code{\link[=seq_nhs_number]{seq_nhs_number()}} \code{\link[=validate_nhs_number]{validate_nhs_number()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}}
}
\author{
Owen Pullen
}
