% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_inpatient_spells.R
\name{cip_spells}
\alias{cip_spells}
\title{Continuous Inpatient Spells}
\usage{
cip_spells(
  .data,
  patient_group_vars,
  spell_start_date,
  admission_method,
  admission_source,
  spell_end_date,
  discharge_destination,
  patient_classification
)
}
\arguments{
\item{.data}{a data frame, tibble; direct or piped (\%>\%)}

\item{patient_group_vars}{a vector containing any variables to be used for record grouping, minimum is a patient identifier}

\item{spell_start_date}{Inpatient provider spell or episode admission date}

\item{admission_method}{CDS admission method code}

\item{admission_source}{CDS admission source code}

\item{spell_end_date}{Inpatient provider spell or episode discharge  date}

\item{discharge_destination}{CDS discharge destination code}

\item{patient_classification}{CDS patient classification code}
}
\value{

}
\description{
A continuous inpatient (CIP) spell is a continuous period of care within the NHS, regardless of any transfers which may take place. It can therefore be made up of one or more provider spells. A CIP spell starts when a decision has been made to admit the patient, and a consultant has taken responsibility for their care. The spell ends when the patient dies or is discharged from hospital
}
\examples{

ex <- dplyr::tribble(
  ~id, ~provider, ~spell_start, ~spell_end, ~adm_meth, ~adm_src, ~dis_meth, ~dis_dest, ~patclass,
  465,    "X1T",  "2020-03-07", "2020-03-07", "21", "19", "1","51",1,
  465,    "X1T",  "2020-03-07", "2020-03-25", "81", "51", "1","51",1,
  465,    "X1T",  "2020-03-25", "2020-04-02", "21", "19", "1","51",1,
  465,    "X1T",  "2020-04-03", "2020-04-27", "81", "51", "1","54",1,
  8418,   "KHA",  "2020-01-25", "2020-01-25", "21", "19", "1","51",1,
  8418,   "KHA",  "2020-01-26", "2020-01-27", "21", "51", "1","19",1,
  8418,   "KHA",  "2020-07-14", "2020-07-17", "11", "19", "1","19",1,
  8418,   "KHA",  "2020-08-02", "2020-08-07", "21", "51", "1","19",1,
  8418,   "KHA",  "2020-08-12", "2020-08-14", "21", "19", "1","19",1,
  8418,   "KHA",  "2020-08-19", "2020-08-19", "21", "19", "1","51",1,
  8418,   "KHA",  "2020-08-19", "2020-08-22", "21", "19", "1","19",1,
  8418,   "KHA",  "2020-11-19", "2020-12-16", "21", "51", "4","79",1,
  26443,  "BX2",  "2019-11-12", "2020-04-17", "21", "19", "1","51",1,
  26443,  "BX2",  "2020-04-17", "2020-04-23", "81", "51", "1","51",1,
  26443,  "BX2",  "2020-04-23", "2020-05-20", "21", "19", "4","79",1,
  33299,  "PXH",  "2020-07-03", "2020-07-24", "81", "51", "1","65",1,
  33299,  "PXH",  "2020-01-17", "2020-01-28", "21", "19", "1","19",1,
  33299,  "PXH",  "2020-02-07", "2020-02-07", "21", "19", "1","19",1,
  33299,  "PXH",  "2020-03-20", "2020-03-23", "21", "19", "1","19",1,
  33299,  "PXH",  "2020-04-27", "2020-04-29", "21", "19", "1","19",1,
  33299,  "PXH",  "2020-06-21", "2020-06-21", "21", "19", "1","19",1,
  33299,  "PXH",  "2020-07-02", "2020-07-03", "21", "19", "1","29",1,
  33299,  "PXH",  "2020-10-17", "2020-11-27", "21", "19", "8","98",1,
  33299,  "PXH",  "2020-11-27", "2021-01-02", "13", "51", "1","51",1,
  33299,  "PXH",  "2021-01-02", "2021-01-10", "13", "19", "4","79",1,
  52635,  "9HA",  "2019-12-31", "2019-12-31", "12", "19", "1","19",2,
  52635,  "9HA",  "2020-01-02", "2020-01-11", "22", "19", "1","19",1,
  52635,  "9HA",  "2020-01-14", "2020-01-14", "12", "19", "1","19",2,
  52635,  "9HA",  "2020-01-16", "2020-02-04", "2D", "19", "1","51",1,
  52635,  "9HA",  "2020-02-07", "2020-02-07", "13", "19", "1","19",2,
  52635,  "9HA",  "2020-02-11", "2020-02-11", "13", "19", "1","19",2,
  52635,  "9HA",  "2020-02-14", "2020-02-14", "13", "19", "1","19",2,
  52635,  "9HA",  "2020-02-18", "2020-02-18", "13", "19", "1","51",2,
  52635,  "9HA",  "2020-02-21", "2020-02-21", "13", "19", "1","51",2,
  52635,  "9HA",  "2020-02-25", "2020-02-25", "13", "19", "1","51",2,
  52635,  "9HA",  "2020-02-28", "2020-02-28", "13", "19", "1","19",2,
  52635,  "9HA",  "2020-03-09", "2020-03-09", "13", "51", "1","19",2,
  52635,  "9HA",  "2020-03-11", "2020-03-11", "13", "51", "1","51",2,
  52635,  "9HA",  "2020-03-12", "2020-03-12", "13", "51", "1","51",2,
  52635,  "9HA",  "2020-03-13", "2020-03-13", "13", "51", "1","19",2,
  52635,  "9HA",  "2020-03-14", "2020-03-30", "21", "19", "1","51",1,
  52635,  "YYT",  "2020-02-04", "2020-02-07", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-07", "2020-02-11", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-11", "2020-02-14", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-14", "2020-02-18", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-18", "2020-02-21", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-21", "2020-02-25", "13", "51", "1","51",1,
  52635,  "YYT",  "2020-02-25", "2020-02-28", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-02-28", "2020-03-09", "81", "51", "1","51",1,
  52635,  "YYT",  "2020-03-09", "2020-03-11", "13", "51", "1","51",1,
  52635,  "YYT",  "2020-03-11", "2020-03-12", "13", "51", "1","51",1,
  52635,  "YYT",  "2020-03-12", "2020-03-13", "13", "51", "1","51",1,
  78915,  "ABX",  "2020-04-16", "2020-04-24", "21", "19", "1","29",1,
  78915,  "ABX",  "2020-04-24", "2020-05-13", "11", "51", "1","54",1,
  78915,  "ABX",  "2020-05-13", "2020-06-11", "81", "51", "2","19",1
) \%>\%
  dplyr::mutate_at(vars(dplyr::contains("spell")),~as.Date(.))

ex \%>\% cip_spells(
  patient_group_vars = c(id,provider),
  patient_classification = patclass,
  spell_start_date = spell_start,
  admission_method = adm_meth,
  admission_source = adm_src,
  spell_end_date = spell_end,
  discharge_destination = dis_dest
)
}
\seealso{
http://content.digital.nhs.uk/media/11859/Provider-Spells-Methodology/pdf/Spells_Methodology.pdf
}
