% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_from_zip.R
\name{csv_from_zip}
\alias{csv_from_zip}
\title{Download a csv from a zip}
\usage{
csv_from_zip(x)
}
\arguments{
\item{x}{a zip file from the web}
}
\value{
a zip file for ingestion into your chosen readr
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
A convenience function to allow you to pull data from NHS, ONS and ODR assets
}
\examples{
\dontrun{
read.csv(csv_from_zip("https://files.digital.nhs.uk/assets/ods/current/succarc.zip"))
}

}
