% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_in_out_dates.R
\name{hospital_in_out_dates}
\alias{hospital_in_out_dates}
\title{Hospital IN/OUT dates}
\usage{
hospital_in_out_dates(
  data,
  person_id = "id",
  hospital = list(org_code = "organisation_code_of_provider", event_date = "ev_date",
    ae_arrive = "arrival_date", ae_depart = "departure_date", ae_discharge =
    "ecds_discharge", in_spell_start = "spell_start_date", in_spell_end =
    "spell_end_date", in_discharge = "discharge_destination")
)
}
\arguments{
\item{data}{the linked asset holding A&E and Inpatient data}

\item{person_id}{the column containing the unique patient ID}

\item{hospital}{a list containing the following items
\describe{
\item{\code{org_code}}{the NHS trust organisation codes}
\item{\code{event_date}}{the comparison date used; often \code{specimen_date}}
\item{\code{ae_arrive}}{the ECDS arrival date}
\item{\code{ae_depart}}{the ECDS discharge date}
\item{\code{ae_discharge}}{the ECDS discharge status; recommend grouping from \code{epidm::lookup_recode}}
\item{\code{in_spell_start}}{the HES/SUS spell start date; recommend after \code{epidm::group_time}}
\item{\code{in_spell_end}}{the HES/SUS spell end date; recommend after \code{epidm::group_time}}
\item{\code{in_discharge}}{the HES/SUS discharge destination code; recommend grouping from \code{epidm::lookup_recode}}
}}
}
\value{
new date columns on the data.table for \code{hospital_in} and \code{hospital_out} and \code{hospital_event_rank}
}
\description{
This function helps to determine when a patient has been in
hospital across  spell aggregation.
When retaining the final record the following criteria is used:
\itemize{
\item{"1"}{Current admissions take priority}
\item{"2"}{When conflicting on the same day, inpatient admissions take priority over A&E emergency care data}
\item{"3"}{Where a patient has a linked A&E admission to a hospital inpatient stay, the A&E admission date is used}
\item{"4"}{Where a patient has a positive test between two hospital stays the most recent completed hospital stay prior to the test is retained except if the time between these events is greater than 14 days, then the first admission following the test is retained}
}
}
\examples{
hospital_in_out_dates(link,
person_id = 'id',
hospital = list(
  org_code = 'organisation_code_of_provider',
  event_date = 'ev_date',
  ae_arrive = 'arrival_date',
  ae_depart = 'departure_date',
  ae_discharge = 'ecds_discharge',
  in_spell_start = 'spell_start_date',
  in_spell_end = 'spell_end_date',
  in_discharge = 'discharge_destination'
)[]

}
\seealso{
epidm::lookup_recode()

epidm::group_time()

epidm::cip_spells()
}
