% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_episode.R
\name{infection_episode}
\alias{infection_episode}
\title{Patient Episode Groupings}
\usage{
infection_episode(
  x,
  group_vars,
  event_date,
  event_window = c(0:Inf),
  indx_name = "indx",
  window_type = c("static", "rolling")
)
}
\arguments{
\item{x}{a data.frame or tibble containing the line list}

\item{group_vars}{a vector containing any columns used to identify the patient, infection, and site}

\item{event_date}{the column as a date variable containing the specimen date}

\item{event_window}{an integer to specify the length of the event_window in time}

\item{indx_name}{a character string to rename an indx column}

\item{window_type}{a choice of static or rolling for how the event_window is calculated}
}
\value{
A new dataframe containing the original data, and columns specifying monomicrobial or polymicrobial patient episodes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Groups single event instances eg. test results in time based on a
static or rolling window in days. Events can be grouped according to a
any number of columns, to group monomicrobial or polymicrobial episodes
eg. group infection episodes within 14-days

this function was deprecated as the full functionality is embedded in group_time()
}
\examples{
episode_test <- structure(
  list(
    pat_id = c(1L, 1L, 1L, 1L, 2L, 2L, 2L,
               1L, 1L, 1L, 1L, 2L, 2L, 2L),
    species = c(rep("E. coli",7),rep("K. pneumonia",7)),
    spec_type = c(rep("Blood",7),rep("Blood",4),rep("Sputum",3)),
    sp_date = structure(c(18262, 18263, 18281, 18282, 18262, 18263, 18281,
                          18265, 18270, 18281, 18283, 18259, 18260, 18281),
                        class = "Date")
  ),
  row.names = c(NA, -14L), class = "data.frame")

infection_episode(
  x = episode_test,
  group_vars = c("pat_id", "spec_type", "species"),
  event_date = "sp_date",
  event_window = 14,
  window_type = 'static',
  indx_name = 'episode_indx'
)[]

}
\seealso{
epidm::group_time
}
\keyword{internal}
