% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_ae_inpatient.R
\name{link_ae_inpatient}
\alias{link_ae_inpatient}
\title{Link A&E to Inpatient records}
\usage{
link_ae_inpatient(
  ae_data,
  ae_arrival_date,
  ae_departure_date,
  inp_data,
  inp_inp_spell_start_date,
  inp_spell_id,
  nhs_number = c("nhs_number", "nhs_number"),
  hospital_number = c("local_patient_identifier", "local_patient_identifier"),
  patient_dob = c("date_birth", "date_birth"),
  org_code = c("organisation_code_of_provider", "organisation_code_code_of_provider"),
  .forceCopy = FALSE
)
}
\arguments{
\item{ae_data}{the ECDS A&E dataset}

\item{ae_arrival_date}{the ECDS arrival date}

\item{inp_data}{the HES/SUS inpatient dataset}

\item{inp_spell_id}{the HES/SUS spell id}

\item{nhs_number}{a vector containing the column names for the NHS numbers in the order \code{c('ae','inpatient')}}

\item{hospital_number}{a vector containing the column names for the Hospital numbers in the order \code{c('ae','inpatient')}}

\item{patient_dob}{a vector containing the column names for the date of birth in the order \code{c('ae','inpatient')}}

\item{org_code}{a vector containing the column names for the organisation codes in the order \code{c('ae','inpatient')}}

\item{.forceCopy}{a boolean to control if you want to copy the dataset before
linking together}

\item{ae_ae_departure_date}{the ECDS discharge date}

\item{admission_date}{a string containing the inpatient (HES/SUS) admission date column name}
}
\value{
a patient level linked hospital record
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Link together ECDS A&E records to HES/SUS inpatient records on
NHS number, Hospital Number and Date of Birth. To note that the inpatient
records should already be aggregated into spells at the desired level.
}
\seealso{
group_time continuous_inpatient_spells
}
