% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_recode.R
\name{lookup_recode}
\alias{lookup_recode}
\title{Lookup table switch handler}
\usage{
lookup_recode(src, type = c("species", "specimen", "manual"), .import = NULL)
}
\arguments{
\item{src}{a character, vector or column containing the value(s) to be referenced}

\item{type}{a character value to denote the lookup table used}

\item{.import}{a list  in the order list(new,old) containing the
values for another lookup table existing in the environment}
}
\value{
a list object of the recoded field
}
\description{
A function to call an epidm lookup table and
recode  where we are aware of a new value.
}
\details{
Built in are  the organism re-classifications and specimen_type groupings
and a manual mode.
#'
}
\examples{
df <- data.frame(
  sp = c(
    sample(respeciate_organism$previous_organism_name,9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  ty = sample(specimen_type_grouping$specimen_type,12),
  dt = sample(seq.Date(from = Sys.Date()-365,
                       to = Sys.Date(),
                       by = "day"),12)
)
df <- df[order(df$dt),]

df$species <- lookup_recode(df$sp,'species')
df$grp <- lookup_recode(df$ty,'specimen')
df

# for a tidyverse use
# df \%>\% mutate(sp=lookup_recode(sp,'species))

lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
}
