% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_inpatient_episode_dates.R
\name{proxy_episode_dates}
\alias{proxy_episode_dates}
\title{HES/SUS Episode Date Cleaning

Correcting for missing end dates on HES/SUS episodes}
\usage{
proxy_episode_dates(
  x,
  patient_group_vars,
  spell_start_date,
  spell_end_date,
  discharge_destination
)
}
\arguments{
\item{x}{a data frame; will be converted to a data.table}

\item{patient_group_vars}{a vector containing any variables to be used for record grouping, minimum is a patient identifier}

\item{spell_start_date}{Inpatient provider spell or episode admission date}

\item{spell_end_date}{Inpatient provider spell or episode discharge  date}

\item{discharge_destination}{CDS discharge destination code}

\item{admission_method}{CDS admission method code}

\item{admission_source}{CDS admission source code}

\item{patient_classification}{CDS patient classification code}
}
\value{
a data.table with cleaned start and end dates, a cip_indx and new cip window dates
}
\description{
HES/SUS Episode Date Cleaning

Correcting for missing end dates on HES/SUS episodes
}
\examples{

cip_test <- data.frame(
  id = c('465','465','465','465','8418','8418','8418','8418','8418','8418','8418','8418','26443','26443','26443','33299','33299','33299','33299','33299','33299','33299','33299','33299','33299','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','52635','78915','78915','78915'),
  provider = c('X1T','X1T','X1T','X1T','KHA','KHA','KHA','KHA','KHA','KHA','KHA','KHA','BX2','BX2','BX2','PXH','PXH','PXH','PXH','PXH','PXH','PXH','PXH','PXH','PXH','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','9HA','YYT','YYT','YYT','YYT','YYT','YYT','YYT','YYT','YYT','YYT','YYT','ABX','ABX','ABX'),
  spell_start = as.Date(c('2020-03-07','2020-03-07','2020-03-25','2020-04-03','2020-01-25','2020-01-26','2020-07-14','2020-08-02','2020-08-12','2020-08-19','2020-08-19','2020-11-19','2019-11-12','2020-04-17','2020-04-23','2020-07-03','2020-01-17','2020-02-07','2020-03-20','2020-04-27','2020-06-21','2020-07-02','2020-10-17','2020-11-27','2021-01-02','2019-12-31','2020-01-02','2020-01-14','2020-01-16','2020-02-07','2020-02-11','2020-02-14','2020-02-18','2020-02-21','2020-02-25','2020-02-28','2020-03-09','2020-03-11','2020-03-12','2020-03-13','2020-03-14','2020-02-04','2020-02-07','2020-02-11','2020-02-14','2020-02-18','2020-02-21','2020-02-25','2020-02-28','2020-03-09','2020-03-11','2020-03-12','2020-04-16','2020-04-24','2020-05-13')),
  spell_end = as.Date(c('2020-03-07','2020-03-25','2020-04-02','2020-04-27','2020-01-25','2020-01-27','2020-07-17','2020-08-07','2020-08-14','2020-08-19','2020-08-22','2020-12-16','2020-04-17','2020-04-23','2020-05-20','2020-07-24','2020-01-28','2020-02-07','2020-03-23','2020-04-29','2020-06-21','2020-07-03','2020-11-27','2021-01-02','2021-01-10','2019-12-31','2020-01-11','2020-01-14','2020-02-04','2020-02-07','2020-02-11','2020-02-14','2020-02-18','2020-02-21','2020-02-25','2020-02-28','2020-03-09','2020-03-11','2020-03-12','2020-03-13','2020-03-30','2020-02-07','2020-02-11','2020-02-14','2020-02-18','2020-02-21','2020-02-25','2020-02-28','2020-03-09','2020-03-11','2020-03-12','2020-03-13','2020-04-24','2020-05-13','2020-06-11')),
  dis_dest = c('51','51','51','54','51','19','19','19','19','51','19','79','51','51','79','65','19','19','19','19','19','29','98','51','79','19','19','19','51','19','19','19','51','51','51','19','19','51','51','19','51','51','51','51','51','51','51','51','51','51','51','51','29','54','19'),
)

proxy_episode_dates(x=cip_test,
  patient_group_vars = c('id','provider'),
  spell_start_date = 'spell_start',
  spell_end_date = 'spell_end',
  discharge_destination = 'dis_dest'
)
}
