% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nhs_number.R
\name{seq_nhs_number}
\alias{seq_nhs_number}
\title{Find sequences in NHS number using regex}
\usage{
seq_nhs_number(value)
}
\arguments{
\item{value}{NHS number find sequences such as 9999999999 in NHS numbers using regex in strings}
}
\description{
Examples are not real NHS numbers please do not use real
NHS numbers, in tests, code or documentation.
Function using regex to remove consecutive 10 digit strings from NHS numbers,
i.e strings contains 10 1's or 0's
variable or string containing numeric NHS numbers as class character
Can be re-purposed to clean other numeric id variables by altering regex {n}
Use: to be used inside dplyr::mutate()
e.g. df \%>\% mutate(nhsnumber = seq_nhs_number(nhsnumber)) # use inside mutate
}
\examples{
nhs_no <- stringr::str_flatten(replicate(10, 9))
seq_nhs_number(nhs_no)
}
\seealso{
\code{\link[=validate_nhs_number]{validate_nhs_number()}} \code{\link[=checksum_nhs_number]{checksum_nhs_number()}} \code{\link[=range_nhs_number]{range_nhs_number()}} \code{\link[=remove_non_numeric_char]{remove_non_numeric_char()}}
}
\author{
Owen Pullen
}
