% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_readwrite.R
\name{sql_connect}
\alias{sql_connect}
\title{Connect to a SQL database}
\usage{
sql_connect(server, database)
}
\arguments{
\item{server}{a string containing the server connection}

\item{database}{a string containing the database name within the data store}
}
\value{
a SQL connection object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

An internal function to help setup connections to SQL databases
acting as a wrapper for the odbc and DBI packages
}
\keyword{internal}
