% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_id.R
\name{uk_patient_id}
\alias{uk_patient_id}
\title{Patient ID record grouping}
\usage{
uk_patient_id(
  x,
  nhs_number,
  hospital_number,
  date_of_birth,
  sex,
  forename = "NONAME",
  surname = "NONAME"
)
}
\arguments{
\item{x}{a data.frame or data.table containing the cleaned line list}

\item{nhs_number}{a column as a character containing the patient NHS numbers}

\item{hospital_number}{a column as a character containing the patient Hospital numbers}

\item{date_of_birth}{a column as a date variable containing the patient date of birth in date format}

\item{sex}{a column as a character containing the patient sex}

\item{forename}{a column as a character containing the patient forename; leave as NONAME if unavailable}

\item{surname}{a column as a character containing the patient surname; leave as NONAME if unavailable}

\item{sort_by}{a column to give a priority sort order if required}
}
\value{
patientID grouping variable addedd to the data.frame

A dataframe with two new variables: id a unique patient id, and n_in_id an integer variable with the number of rows in the id
}
\description{
Groups patient records from multiple isolates with a single patientID
GROUPED by nhs_number, hospital_number, date_of_birth, assigns a unique IDno to each patient.
sort by specimen date
generates a unique sequential number from 1-X for each record
then based on grouping, reassigns records to same number
group by different combos of patient identifiers, also dont include invalid nhs_numbers

Grouping is based on five stages:
\enumerate{
\item matching nhs number and date of birth
\item Hospital number &  Date of Birth
\item NHS number & Hospital Number
\item Sex & Date of Birth & Surname IF nhs unknown
\item Sex & Date of Birth & Fuzzy Name
}
}
\examples{
dat <- data.frame(
  nhs_n = c(
    9434765919,9434765919,9434765919,NA,NA,
    3367170666,5185293519,5185293519,5185293519,8082318562,NA,NA,NA
  ),
  hosp_n = c(
    '13','13','13','NA','13','13','13','31','31','96','96','96','96'
  ),
  sex = c('M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'F', 'U', 'U', 'F'),
  dateofbirth = as.Date(
    c(
      '1988-10-06','1988-06-10','1900-01-01','1988-10-06','1988-10-06',
      '1988-10-06','1988-10-06','1988-10-06','1988-10-06','2020-01-28',
      '2020-01-28','2020-01-28','2020-01-28'
    )
  ),
  firstname = c(
    'Danger','Danger','Denger','Danger','Danger','DANGER','Danger',
    'Danger','Danger','Crazy','Crazy','Krazy','C'
  ),
  lastname = c(
    'Mouse','Mause','Mouse','Moose','Moose','Mouse','MOUSe',
    'Mouse','Mouse','Frog','FROG','Frug','Frog'
  )
)

}
