% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_id.R
\name{uk_patient_id}
\alias{uk_patient_id}
\title{Patient ID record grouping}
\usage{
uk_patient_id(
  x,
  nhs_number,
  hospital_number,
  date_of_birth,
  sex,
  forename = "NONAME",
  surname = "NONAME"
)
}
\arguments{
\item{x}{a data.frame or data.table containing the cleaned line list}

\item{nhs_number}{a column as a character containing the patient NHS numbers}

\item{hospital_number}{a column as a character containing the patient Hospital numbers}

\item{date_of_birth}{a column as a date variable containing the patient date of birth in date format}

\item{sex}{a column as a character containing the patient sex}

\item{forename}{a column as a character containing the patient forename; leave as NONAME if unavailable}

\item{surname}{a column as a character containing the patient surname; leave as NONAME if unavailable}

\item{sort_by}{a column to give a priority sort order if required}
}
\value{
patientID grouping variable addedd to the data.frame

A dataframe with two new variables: id a unique patient id, and n_in_id an integer variable with the number of rows in the id
}
\description{
Groups patient records from multiple isolates with a single integer patientID
by grouping patient identifiers.

Grouping is based on five stages:
\enumerate{
\item matching nhs number and date of birth
\item Hospital number &  Date of Birth
\item NHS number & Hospital Number
\item Date of Birth & Surname IF nhs unknown
\item Sex & Date of Birth & Fuzzy Name
}
}
\examples{
id_test <- data.frame(
  nhs_n = c(
    9434765919,9434765919,9434765919,NA,NA,
    3367170666,5185293519,5185293519,5185293519,8082318562,NA,NA,NA
  ),
  hosp_n = c(
    '13','13','13','UNKNOWN','13','13','13','31','31','96','96',NA,'96',
  ),
  sex = c('M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'F', 'U', 'U', 'F'),
  dateofbirth = as.Date(
    c(
      '1988-10-06','1988-06-10','1900-01-01','1988-10-06','1988-10-06',
      '1988-10-06','1988-10-06','1988-10-06','1988-10-06','2020-01-28',
      '2020-01-28','2020-01-28','2020-01-28'
    )
  ),
  firstname = c(
    'Danger','Danger','Denger','Danger','Danger','DANGER','Danger',
    'Danger','Danger','Crazy','Crazy','Krazy','C'
  ),
  lastname = c(
    'Mouse','Mause','Mouse','Moose','Moose','Mouse','MOUSe',
    'Mouse','Mouse','Frog','FROG','Frug','Frog'
  )
)
id_test <- uk_patient_id(x = id_test,
                   nhs_number = 'nhs_n',
                   hospital_number = 'hosp_n',
                   forename = 'firstname',
                   surname = 'lastname',
                   sex = 'sex',
                   date_of_birth = 'dateofbirth')
print(id_test)

}
